requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
build==1.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Cython==3.0.7
discord.py==2.3.2
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.5
packaging==23.2
pillow==10.2.0
pyproject_hooks==1.0.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.24
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.1
yarl==1.9.4